Switched 2D drawing library from Skia to Cairo.
This decision was made due to the difficulties associated with integrating
a Skia build into the CMake build and the uncertainties associated with
Skia's ill-documented build system. Cairo was chosen due to its close
association with GStreamer and the high probability of it being available
on systems that have GStreamer.

Some slowdown does occur, though this appears to mostly be due to better
antialiasing which is enabled by default in Cairo.